refactor(application): extract resolution_guide_builder and upgrade_recommendation_builder
Extract build_resolution_guide, build_resolution_entry_view, and
build_upgrade_recommendations from sbom_read_model_builder/mod.rs into
dedicated sub-builder files as the final step of #314.

- Create resolution_guide_builder.rs with build_resolution_guide (pub(super))
  and build_resolution_entry_view (private)
- Create upgrade_recommendation_builder.rs with build_upgrade_recommendations (pub(super))
- Update mod.rs to delegate to new sub-builders and remove extracted methods
- mod.rs now contains only orchestrator logic (~80 lines, excluding test module)

Closes #341
